Fix plan phases shown differently across the plugin
